+++ mpv.net/Misc/Extension.cs
@@ -4,6 +4,7 @@ using System.ComponentModel.Composition;
 using System.ComponentModel.Composition.Hosting;
 using System.IO;
 using System.Windows.Forms;
+using System.Linq;
 
 namespace mpvnet
 {
@@ -19,12 +20,20 @@ namespace mpvnet
             try
             {
                 AggregateCatalog catalog = new AggregateCatalog();
-
                 string dir = Application.StartupPath + "\\Extensions";
 
                 if (Directory.Exists(dir))
-                    foreach (string i in Directory.GetDirectories(dir))
-                        catalog.Catalogs.Add(new DirectoryCatalog(i, "*Extension.dll"));
+                {
+                    string[] knownExtensions = { "RatingExtension", "ScriptingExtension" };
+
+                    foreach (string path in Directory.GetDirectories(dir))
+                    {
+                        if (knownExtensions.Contains(Path.GetFileName(path)))
+                            catalog.Catalogs.Add(new DirectoryCatalog(path, "*Extension.dll"));
+                        else
+                            Msg.ShowError("Failed to load extension", "Only extensions that ship with mpv.net are allowed in <startup>\\extensions.\n\nUser extensions have to use <config folder>\\extensions.\n\nNever copy a new mpv.net version over a old mpv.net version.\n\nNever install a new mpv.net version on top of a old mpv.net version.\n\n" + path);
+                    }
+                }
 
                 dir = mp.ConfigFolder + "\\Extensions";